requirements.txt
# Core dependencies
opencv-python>=4.8.0
pillow
pytesseract
numpy
tqdm

# Web Interface
flask>=2.3.0

# ML and Processing
tensorflow>=2.12.0
imagehash
requests

# Modern Face Detection (highly recommended)
mediapipe>=0.10.0

# Optional Advanced Face Recognition (can be tricky to install)
# face_recognition
# dlib

# If face_recognition installation fails:
# 1. Install Visual Studio Build Tools or Visual Studio Community
# 2. Install CMake: https://cmake.org/download/
# 3. pip install dlib
# 4. pip install face_recognition

# Alternative lightweight options:
# tensorflow-lite-runtime  # Use this instead of tensorflow for smaller installs
# onnxruntime             # Alternative to TensorFlow

# Built-in packages (no installation needed):
# - sqlite3
# - hashlib
# - pathlib
# - datetime
# - logging
